pyproject.toml
[project]
name = "r2x-plexos"
version = "0.1.3"
description = "Plexos Models, Parser and Exporter"
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "plexosdb>=1.3.0",
    "r2x-core>=0.4.0,<1.0.0",
]
keywords=["PLEXOS"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: File Formats :: JSON",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[dependency-groups]
dev = [
    "ipython>=9.2.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-coverage>=0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.11.5",
]
docs = [
    "autodoc-pydantic",
    "docstr-coverage>=2.3.2",
    "furo>=2025.9.25",
    "ghp-import",
    "myst-parser",
    "sphinx>=8.0.0",
    "sphinx-copybutton",
    "sphinx-reports",
    "sphinx-tabs",
]

[project.entry-points."r2x_plugin"]
plexos-parser = "r2x_plexos:PLEXOSParser"
plexos-exporter = "r2x_plexos:PLEXOSExporter"

[project.urls]
Homepage = "https://github.com/NREL/r2x-plexos"
Documentation = "https://nrel.github.io/r2x-plexos/"
Repository = "https://github.com/NREL/r2x-plexos"
Issues = "https://github.com/NREL/r2x-plexos/issues"
Changelog = "https://nrel.github.io/r2x-plexos/CHANGELOG.html"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = [
    "--cov=r2x_plexos",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=json",
    "--strict-markers",
    "-v",
]
markers = [
    "unit: Fast unit tests with no external dependencies",
    "integration: Integration tests requiring system setup",
    "slow: Slow tests (>1s)",
    "export: marks to test exporter functions",
    "serialize: marks tests for serialization functionality",
    "fixtures: marks for testing fixtures"
]


[tool.mypy]
strict = true
exclude = ["tests/"]
plugins = [
  "pydantic.mypy"
]


[tool.ruff]
line-length = 110
extend-exclude = [
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "D",      # pyddoc
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PIE",    # flake8-pie
    "PERF",   # perflint (performance)
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B027",   # Empty method in abstract class without @abstractmethod (intentional for hooks)
]


[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
# Ignore `D` rules everywhere except for the `src/` directory.
"!src/**.py" = ["D", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["r2x_core"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.mccabe]
# Cyclomatic complexity threshold
max-complexity = 10

[tool.coverage.run]
source = ["src/r2x_plexos"]
omit = [
  "*/tests/*",
  "*/__pycache__/*",
  "*/site-packages/*",
  "src/r2x_plexos/plugins.py",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
  "@abstract",
]
precision = 2
fail_under = 90

[tool.coverage.html]
directory = "htmlcov"
